refactor(#385): move SpdlogLogger to separate header — decouple ilogger.h from spdlog
- Move SpdlogLogger class to spdlog_logger.h (new file)
- ilogger.h no longer includes spdlog headers — uses fmt directly
- Add StderrFallbackLogger as pre-init default (zero spdlog dep)
- log_config.h installs SpdlogLogger via set_logger() at init
- Add explicit fmt::fmt CMake dependency (was transitive via spdlog)
- Fix transitive include deps: mission_fsm.h <vector>, gimbal_controller.h <algorithm>
