+++ CMakeLists.txt
@@ -85,6 +85,7 @@ endif()
 
 # ── Find system dependencies ────────────────────────────────
 find_package(Threads       REQUIRED)
+find_package(fmt           REQUIRED)
 find_package(spdlog        REQUIRED)
 find_package(Eigen3        REQUIRED NO_MODULE)
 find_package(nlohmann_json REQUIRED)
