+++ process4_mission_planner/include/planner/mission_fsm.h
@@ -6,6 +6,7 @@
 
 #include <array>
 #include <string>
+#include <vector>
 
 namespace drone::planner {
